+++ .env.sample
@@ -1,0 +1,1 @@
+API_KEY=discord_api_key
