+++ .gitignore
@@ -139,3 +139,6 @@ cython_debug/
 
 # Visual Studio code
 .vscode/
+
+# Environment variables file
+.env
