+++ README.md
@@ -1,0 +1,17 @@
+# ChancelerPalpatine
+
+I love democracy! And Python!
+
+## Setup
+
+```bash
+python3 -m virtualenv venv
+source venv/bin/activate
+pip3 install -r requirements.txt
+```
+
+## Updating dependencies
+
+```bash
+pip3 freeze > requirements.txt
+```
