+++ bot/level.py
@@ -3,7 +3,7 @@ import random
 import json
 import os
 from discord.ext import commands
-from client import client
+from bot import client
 
 @client.event
 async def on_member_join(member):
