--- python/run.py
@@ -1,6 +1,0 @@
-import discord
-from client import client
-from bot import *
-from level import *
-
-client.run('NzUxMDg3NjY1OTUyMDYzNTgw.X1D-5g.mFT9Vzes5CvYn94CjX_VoMIMY98')
