--- python/users.json
@@ -1,3 +1,0 @@
-{
-  
-}
